Car_rental/requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.2.19
django-paypal==2.1
idna==3.10
numpy==1.24.4
pandas==2.0.3
paypalrestsdk==1.13.3
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
reportlab==4.3.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
XlsxWriter==3.2.2